extract BaseHelper from ActiveRecord spec helpers.
