+++ spec/support/active_record/sqlite3_helper.rb
@@ -1,20 +1,7 @@
-require 'support/active_record/database_setup'
-require 'support/active_record/schema_setup'
+require 'support/active_record/base_helper'
 
-class SQLite3Helper
+class SQLite3Helper < BaseHelper
   puts "Active Record #{ActiveRecord::VERSION::STRING}, sqlite3"
-
-  # ActiveRecord::Base.logger = Logger.new(STDERR)
-
-  def setup
-    create_db
-    establish_connection
-    active_record_load_schema
-  end
-
-  def connection
-    ActiveRecord::Base.connection
-  end
 
   def teardown
     ActiveRecord::Base.connection.truncate_table('users')
@@ -30,10 +17,6 @@ class SQLite3Helper
   def create_db
     @encoding = default_config['encoding'] || ENV['CHARSET'] || 'utf8'
     establish_connection(default_config.merge('database' => 'sqlite3', 'schema_search_path' => 'public'))
-  end
-
-  def establish_connection(config = default_config)
-    ActiveRecord::Base.establish_connection(config)
   end
 end
